backend/requirements/base.txt
Django==3.2
djangorestframework
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-cors-headers==3.10.0
django-environ==0.8.1
pytest==6.2.4
pytest-django==4.5.2